requirement.txt
Django~=3.2.4
djangorestframework~=3.12.4
django-environ~=0.4.5
django-filter~=2.4.0
